requirements.txt
anyio==4.0.0
build==1.0.3
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cryptography==41.0.4
cthenp @ file:///home/lrq619/proj/CTP/dist/cthenp-0.1.2-py3-none-any.whl#sha256=9ac57ec5f403749b278108c33499eb531c710b5fcc1839490b4cfb348a99288b
distlib==0.3.6
docutils==0.20.1
editables==0.5
exceptiongroup==1.1.3
filelock==3.12.0
futures==3.0.5
google-api-core==2.12.0
google-api-python-client==2.103.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
googleapis-common-protos==1.61.0
grpcio==1.51.1
grpcio-tools==1.44.0
h11==0.14.0
hatch==1.7.0
hatchling==1.18.0
httpcore==1.0.1
httplib2==0.22.0
httpx==0.25.1
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
iniconfig==2.0.0
jaraco.classes==3.3.0
jeepney==0.8.0
keyring==24.2.0
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==10.1.0
nh3==0.2.14
packaging==23.2
pathspec==0.11.2
pexpect==4.8.0
pkginfo==1.9.6
platformdirs==3.5.1
pluggy==1.3.0
protobuf==3.20.3
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.16.1
pyparsing==3.1.1
pyperclip==1.8.2
pyproject_hooks==1.0.0
pytest==7.4.2
readme-renderer==42.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.6.0
rsa==4.9
SecretStorage==3.3.3
shellingham==1.5.4
sniffio==1.3.0
tomli==2.0.1
tomli_w==1.0.0
tomlkit==0.12.2
trove-classifiers==2023.11.7
twine==4.0.2
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.6
userpath==1.9.1
virtualenv==20.23.0
zipp==3.17.0
